removed layer_tools since it was causing issues
